init: docs and refactor

+++ src/core/helper/driver.h
@@ -16,19 +16,19 @@ struct driver_t {
     /* geometry */
     uint32_t sector_size;
     uint64_t total_size_bytes;
-    uint32_t total_sectors;
+    uint64_t total_sectors;
 
     /* lifecycle */
     int  (*init)(driver_t *self);
     void (*deinit)(driver_t *self);
 
     /* I/O */
-    int (*read_block)(driver_t *self, uint32_t lba, uint8_t *buf);
-    int (*write_block)(driver_t *self, uint32_t lba, const uint8_t *buf);
+    int (*read_block)(driver_t *self, uint64_t lba, uint8_t *buf);
+    int (*write_block)(driver_t *self, uint64_t lba, const uint8_t *buf);
 
     /* optional multi-block (may be NULL) */
-    int (*read_blocks)(driver_t *self, uint32_t lba, uint8_t *buf, uint32_t count);
-    int (*write_blocks)(driver_t *self, uint32_t lba, const uint8_t *buf, uint32_t count);
+    int (*read_blocks)(driver_t *self, uint64_t lba, uint8_t *buf, uint32_t count);
+    int (*write_blocks)(driver_t *self, uint64_t lba, const uint8_t *buf, uint32_t count);
 
     /* optional flush (may be NULL) */
     int (*sync)(driver_t *self);
